Separate JSON mapping from event log data classes CIRC-1025
