+++ src/test/java/api/support/data/events/log/CheckOutLogEvent.java
@@ -1,0 +1,17 @@
+package api.support.data.events.log;
+
+import java.util.List;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+
+@AllArgsConstructor
+@Builder
+public class CheckOutLogEvent {
+  public final String loanId;
+  public final List<ChangedRequest> changedRequests;
+
+  public ChangedRequest firstChangedRequest() {
+    return changedRequests.get(0);
+  }
+}
