+++ src/test/java/api/support/data/events/log/JsonToCheckOutLogEventMapper.java
@@ -1,4 +1,4 @@
-package api.support.dto;
+package api.support.data.events.log;
 
 import static org.folio.circulation.support.json.JsonObjectArrayPropertyFetcher.mapToList;
 import static org.folio.circulation.support.json.JsonPropertyFetcher.getProperty;
@@ -6,27 +6,16 @@ import static org.folio.circulation.support.json.JsonPropertyFetcher.getProperty
 import java.util.List;
 
 import io.vertx.core.json.JsonObject;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
 
-@AllArgsConstructor
-@Builder
-public class CheckOutLogEvent {
-  public final String loanId;
-  public final List<ChangedRequest> changedRequests;
-
-  public ChangedRequest firstChangedRequest() {
-    return changedRequests.get(0);
-  }
-
-  public static CheckOutLogEvent fromJson(JsonObject json) {
+public class JsonToCheckOutLogEventMapper {
+  public CheckOutLogEvent fromJson(JsonObject json) {
     return CheckOutLogEvent.builder()
       .loanId(getProperty(json, "loanId"))
       .changedRequests(changedRequestsFromJson(json))
       .build();
   }
 
-  private static List<ChangedRequest> changedRequestsFromJson(JsonObject json) {
+  private List<ChangedRequest> changedRequestsFromJson(JsonObject json) {
     return mapToList(json, "requests", ChangedRequest::fromJson);
   }
 }
